Debugs & wait error mitigation added (works)
Been getting a random error from the wait command that happens once every 3-8 hours.  So I added some debug prints and instead of breaking from the loop when it happens, I decided to continue the loop instead and thereby drop a frame.  It's running now to see what happens in a few hours.