Resolved issue #12. The duration of the recorded video is no longer used to determine when the end of the experiment will be, because with a variable frame rate, the duration is inaccurate.
The askFPS default was changed to 0 because of the fact that the frame rate is meaningless given the variable nature of the video recording.

The default writeToFileTiming value was changed to 20.

Attempted to decrease analysis time by reducing the pause in the loop when processing a saved file. This starts to address issue #13. The pause could not be taken out because the figure would not display without it. I'm not sure whether reducing the pause time is making the video processing any faster. The speed may be limited by the live playback.
\nFiles checked in: flyMultiPlateScript.m