Added the ability to select 1 of 2 plate types: the original 96 well format and a new 24 well format.
I added a method called setmarkersize so that the circles drawn over the plate image are relative in size to the actual wells in the image. The marker circle sizes should correspond to the well sizes in the image and it should change as the image/figure is resized.

Files checked in: repositionCrosses.m