Un-flipped the wells so that they represent the actual orientation of the raw video.
I confirmed that this was the correct orientation by using landmarks on the plates that are outside the wells (the feet that the plate rests on that show up as dark circles). By removing the flipping code, the processed image now looks contiguously representative of what the plate actually looks like.

Files checked in: flyMultiPlateScript.m